requirements.txt
pip~=20.2.3
pytz~=2025.2
certifi~=2025.8.3
tornado~=6.1
tzlocal~=5.3.1
python-telegram-bot~=13.15
APScheduler~=3.6.3
cachetools~=4.2.2
six~=1.17.0
setuptools~=49.2.1
python-dotenv~=1.1.1